backend/test_google_ai_latest.py
import os
from dotenv import load_dotenv
import google.generativeai as genai

load_dotenv()

api_key = os.getenv("GOOGLE_API_KEY")
genai.configure(api_key=api_key)

print("Testing gemini-flash-latest...")
try:
    model = genai.GenerativeModel('gemini-flash-latest')
    response = model.generate_content("Ciao!")
    print(f"Success! Response: {response.text}")
except Exception as e:
    print(f"Error with gemini-flash-latest: {e}")
