backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
requests
garminconnect
psycopg2-binary
bcrypt
passlib[bcrypt]
python-multipart
python-jose[cryptography]
httpx
email-validator
openai
google-generativeai
